+++ CHANGELOG.md
@@ -2,12 +2,15 @@
 
 ## [Unreleased]
 ### Added
-- _body attribute containing raw response text on query results
-- SchedulingBooked.Query
 
 ### Changed
 
 ### Removed
+
+## [0.3.0] - 2020-08-27
+### Added
+- _body attribute containing raw response text on query results
+- SchedulingBooked.Query
 
 ## [0.2.0] - 2020-08-16
 ### Added
@@ -22,4 +25,5 @@
 
 [0.1.0]: https://github.com/patient-discovery/redox-client/releases/tag/v0.1.0
 [0.2.0]: https://github.com/patient-discovery/redox-client/releases/tag/v0.2.0
-[Unreleased]: https://github.com/patient-discovery/redox-client/compare/v0.2.0...HEAD
+[0.3.0]: https://github.com/patient-discovery/redox-client/releases/tag/v0.3.0
+[Unreleased]: https://github.com/patient-discovery/redox-client/compare/v0.3.0...HEAD